# Conflicts:
#	src/1. Layers/1.1 Presentation/Well.Api/Controllers/WidgetController.cs
#	src/1. Layers/1.1 Presentation/Well.Api/DependencyResolution/DefaultRegistry.cs
#	src/1. Layers/1.1 Presentation/Well.Dashboard/app/appModule.ts
#	src/1. Layers/1.4 Infrastructure/Well.Repositories/StoredProcedures.cs
#	src/3. Database/Well/Well.sqlproj

+++ src/1. Layers/1.3 Domain/Well.Domain/CreditThreshold.cs
@@ -15,7 +15,7 @@
 
         public int ThresholdLevelId { get; set; }
 
-        public int Threshold { get; set; }
+        public decimal Threshold { get; set; }
 
         public Collection<Branch> Branches { get; set; }
     }
